Continue protocol implementation.

+++ hermes-five/src/protocols/i2c_reply.rs
@@ -2,7 +2,7 @@
 #[derive(Default, Debug, Clone)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct I2CReply {
-    pub address: i32,
-    pub register: i32,
-    pub data: Vec<u8>,
+    pub address: u16,
+    pub register: u16,
+    pub data: Vec<u16>,
 }
